% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ABCGCB.R
\name{ABCGCB}
\alias{ABCGCB}
\alias{ABCGCBClassifier}
\alias{ABCGCBclassifier}
\title{Classification of cell of origin}
\usage{
ABCGCB(new.data, NC.range = c(0.1, 0.9))
}
\arguments{
\item{new.data}{An expression matrix.}

\item{NC.range}{A numeric vector with values for which the probabilities should be cut.}
}
\value{
Probabilities regarding each patients association with each class.
}
\description{
Classification according to ABC/GCB.
}
\details{
The function ABC/GCB classifies DLBCL patients according to the cell of origin for the tumor.
}
\examples{
require(affy)
require(MATT)
u133 <- exprs(readRDS(system.file("extdata/GEPhgu.rda", package = "MATT")))

u133 <- microarrayScale(u133)

ABCGCB(u133)


huex <- exprs(readRDS(system.file("extdata/GEPexon.full.rda",
                                  package = "MATT")))

u133.conv  <- huex10st_to_u133(huex, method = "weighted", type = "Complex", min.pct = 0.95)
ABCGCB(u133.conv)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the ABC/GCB and hemaClass paper.
}

