% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAGS2clinic.R
\name{BAGS2Clinic}
\alias{BAGS2Clinic}
\alias{BAGS2ClinicClassifier}
\alias{BAGS2Clinicclassifier}
\title{Classification of cell of origin with Nanostring data}
\usage{
BAGS2Clinic(new.data, cut.spec = NULL, percent.classified = 85)
}
\arguments{
\item{new.data}{An expression \code{matrix} containing the normalized data, 
with samples in columns and genes in rows. Rownames should be genenames in 
hgnc nomenclature.}

\item{cut.spec}{A \code{numeric} value used to threshold the probabilities 
and determine the class.}

\item{percent.classified}{A \code{numeric} value indicating the percentage of 
patients that should be classified. An alternative to \code{cut.spec}.}
}
\value{
A \code{list} of probabilities regarding each patients association 
  with each class, the determined class, and the used cut-off thresholds.
}
\description{
Classification using the B-cell associated gene signatures (BAGS) classifier 
of Dybkaer et al. [1] adjusted for nanostring data by Michaelsen et al. [2].
The BAGS classifier attemps to assign the given 
supplied Diffuse Large B-Cell Lymphomas (DLBCL) 
samples to one of four classes according to their resembalace to normal 
Centrocytes (CC), Centroblasts (CB), Memory (M), and 
Plasmablasts (PB) cells.
}
\details{
The function BAGS2Clinic classifies DLBCL patients according to the cell 
  of origin for the tumor [1,2].
}
\examples{
\donttest{
# Nanostring files may be read and normalized using the NanoStringNorm package
library(NanoStringNorm)
data.nano.raw <- read.markup.RCC("path to files")

data.nano.norm <- NanoStringNorm(data.nano.raw, 
                                 Background = "mean",
                                 CodeCount = "sum",
                                 return.matrix.of.endogenous.probes = FALSE)

# Data should be log2 transformed and scaled
data.nano.sc <- microarrayScale(log2(as.matrix(data.nano.norm$normalized.data[,-c(1,2,3)]) + 1))
 
# We may now use the BAGS classifier
BAGS2Clinic(data.nano.sc)
}                                                                                                                  
}
\references{
[1] Dybkaer, Karen, Martin Boegsted, Steffen Falgreen, Julie S. Boedker, 
Malene K. Kjeldsen, Alexander Schmitz, Anders E. Bilgrau et al. 
"Diffuse large B-cell lymphoma classification system that associates 
normal B-cell subset phenotypes with prognosis." 
Journal of Clinical Oncology 33, no. 12 (2015): 1379-1388.

[2] Michaelsen, T. Y. et al. "A B-cell-associated gene signature
 classification of diffuse large B-cell lymphoma by NanoString
  technology." Blood advances, 2(13) (2018): 1542-1546.
}
\author{
Thomas Yssing Michaelsen <tym (at) bio.aau.dk> \cr 
  Rasmus Broendum <rfb (at) rn.dk>
}
