% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cov2cor}
\alias{cov2cor}
\title{A conversion from covariance to correlation matrix}
\usage{
cov2cor(S)
}
\arguments{
\item{S}{A covariance matrix}
}
\value{
A correlation matrix of the same size as \code{S}
}
\description{
A conversion from covariance to correlation matrix
}
\details{
Some more details
}
\note{
Some note and warnings of use.
}
\examples{
X <- replicate(5, rnorm(10))
S <- cov(X)
cov2cor(S)
}
\author{
Steffen Falgreen Larsen
  Anders Ellern Bilgrau
}

