% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fitOS.R
\name{fitOS}
\alias{fitOS}
\alias{fitPFS}
\title{Fit overall and progression free survival}
\usage{
fitOS()

fitPFS()
}
\value{
An object of class \code{coxph}, see \code{\link[survival]{coxph}}.
}
\description{
Internal functions to fit overall survival and progression free survival
from IDRC, LLMPP, and MDFCI data.
}
\examples{
hemaClass:::fitOS()
hemaClass:::fitPFS()
}
\keyword{internal}

