% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{hemaClass-package}
\alias{HEMACLASS}
\alias{hemaClass-package}
\alias{hemaclass}
\title{hemaClass: Classification tool for haematological cancers}
\description{
An R-package for DNA micro array classification.
}
\examples{
# See the following for all exported functions
ls("package:hemaClass")
}
\author{
Steffen Falgreen Larsen <sfl (at) rn.dk> \cr
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

