% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeAll.R
\name{makeAll}
\alias{makeAll}
\alias{makeCelfiles}
\alias{makeHowto}
\alias{makeNews}
\title{Generate news, celfiles, and help for website}
\usage{
makeAll()

makeNews(indir = "inst", outdir = "inst/website/www")

makeCelfiles(outdir = "inst/website/www")

makeHowto(indir = "vignettes", outdir = "inst/website/www")
}
\arguments{
\item{indir,}{outdir The in and output directories to read and write the
files from and to.}

\item{overwrite}{Overwrite existing files.}
}
\value{
Writes to files in \code{outdir}. Invisibly returns the path to the
  written file.
}
\description{
Functions for generating the \code{News.html}, \code{celfiles.zip}, and
\code{howto.html} files for the website from \code{NEWS.Rd}, the celfiles in
\code{"/"}, an \code{vignettes/howto.Rmd}. Intented mostly for internal use.
}
\note{
The functions assumes by default that the working directory is the
  root of the local \pkg{hemaClass} package. \cr
  The functions overwrites all previously present files.
}
\examples{
\dontrun{
makeAll()

# makeAll is an alias for:
makeNews()
#makeCelfiles()
makeHowto()
}
}
\author{
Anders Ellern Bilgrau
}

