% Generated by roxygen2 (4.0.2): do not edit by hand
\name{microarrayScale}
\alias{microarrayScale}
\title{Scaling of expression values}
\usage{
microarrayScale(x)
}
\arguments{
\item{x}{matrix containing expression values. May also be of class ExpressionSet.}
}
\value{
Expression matrix consisting of genewise median centred and scaled genes.
}
\description{
Scale expression values prior to use of classification algorithms.
}
\details{
First each probe-set is median centred. Next, each probe-set is scaled to have unit-variance.
}
\examples{
2+2
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass.com paper
}

