% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ABCGCBCoef.R
\name{readABCGCBCoef}
\alias{readABCGCBCoef}
\title{Read coefficients for the ABC/GCB classifier}
\usage{
readABCGCBCoef()
}
\value{
A \code{matrix} with the coefficients.
}
\description{
Function to read or get the coefficients used in the linear predictor score
of the ABC/GCB classifier.
}
\examples{
coefs <- readABCGCBCoef()
head(coefs)
str(coefs)
}

