% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBAGS2clinicCoef.R, R/readBAGSCoef.R
\name{readBAGS2clinicCoef}
\alias{readBAGS2clinicCoef}
\alias{readBAGSCoef}
\title{Read coefficients for the BAGS2clinic classifier}
\usage{
readBAGS2clinicCoef()

readBAGSCoef()
}
\value{
A matrix with coefficients

A matrix with coefficients
}
\description{
Function to read coefficients for the BAGS2clinic classifier.

Function to read coefficients for the BAGS classifier.
}
\examples{
head(readBAGS2clinicCoef())
head(readBAGSCoef())
}
