% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDrugCoef.R
\name{readPredCoef}
\alias{readPredCoef}
\alias{readDrugCoef}
\alias{readClasCoef}
\alias{readCyclophosphamideClasCoef}
\alias{readDoxorubicinClasCoef}
\alias{readVincristineClasCoef}
\alias{readCyclophosphamidePredCoef}
\alias{readDoxorubicinPredCoef}
\alias{readVincristinePredCoef}
\alias{readRituximabClasCoef}
\alias{readRituximabPredCoef}
\alias{readRituximabClasCorCoef}
\alias{readRituximabPredCorCoef}
\alias{readRituximabClasLytiskCoef}
\alias{readDexamethasoneClasCoef}
\alias{readDexamethasonePredCoef}
\alias{readMelphalanClasCoef}
\alias{readMelphalanPredCoef}
\title{Read coefficients for REGS classifiers and predictors}
\usage{
readPredCoef()

readClasCoef()

readCyclophosphamideClasCoef()

readCyclophosphamidePredCoef()

readDoxorubicinClasCoef()

readDoxorubicinPredCoef()

readVincristineClasCoef()

readVincristinePredCoef()

readRituximabClasCoef()

readRituximabPredCoef()

readRituximabClasCorCoef()

readRituximabPredCorCoef()

readRituximabClasLytiskCoef()

readDexamethasoneClasCoef()

readDexamethasonePredCoef()

readMelphalanClasCoef()

readMelphalanPredCoef()
}
\value{
A \code{matrix} of coefficients.
}
\description{
Functions to read coefficients for REGS classifiers and predictors.
}
\examples{
head(readPredCoef())
head(readClasCoef())
head(readCyclophosphamideClasCoef())
head(readCyclophosphamidePredCoef())
head(readDoxorubicinClasCoef())
head(readDoxorubicinPredCoef())
head(readVincristineClasCoef())
head(readVincristinePredCoef())
head(readRituximabClasCoef())
head(readRituximabPredCoef())
head(readRituximabClasCorCoef())
head(readRituximabPredCorCoef())
head(readRituximabClasLytiskCoef())
head(readDexamethasoneClasCoef())
head(readDexamethasonePredCoef())
head(readMelphalanClasCoef())
head(readMelphalanPredCoef())
}
