% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/REGS.R
\name{ResistanceClassifier}
\alias{CyclophosphamideClassifier}
\alias{CyclophosphamidePredictor}
\alias{DexamethasoneClassifier}
\alias{DexamethasonePredictor}
\alias{DoxorubicinClassifier}
\alias{DoxorubicinPredictor}
\alias{MelphalanClassifier}
\alias{MelphalanPredictor}
\alias{REGS}
\alias{ResistanceClassifier}
\alias{ResistancePredictor}
\alias{RituximabClassifier}
\alias{RituximabPredictor}
\alias{VincristineClassifier}
\alias{VincristinePredictor}
\title{REGS for prediction of chemoresistance}
\usage{
ResistanceClassifier(new.data, drugs = c("Cyclophosphamide", "Doxorubicin",
  "Vincristine", "Combined"), cut = list(Cyclophosphamide = c(0.455, 0.67),
  Doxorubicin = c(0.1, 0.86), Vincristine = c(0.38, 0.54), Combined = c(0.067,
  0.907)))

ResistancePredictor(new.data, drugs = c("Cyclophosphamide", "Doxorubicin",
  "Vincristine", "Combined"), cut = list(Cyclophosphamide = c(280, 340),
  Doxorubicin = c(280, 320), Vincristine = c(115, 127), Combined = c(200, 295)))

CyclophosphamideClassifier(new.data)

CyclophosphamidePredictor(new.data)

DoxorubicinClassifier(new.data)

DoxorubicinPredictor(new.data)

VincristineClassifier(new.data)

VincristinePredictor(new.data)

RituximabClassifier(new.data, type = "corrected", cut = c(0.33, 0.66),
  calc.cut = NULL, cut.spec = NULL, percent.classified = 85)

RituximabPredictor(new.data, type = "corrected", cut = c(0.33, 0.66),
  calc.cut = NULL)

DexamethasoneClassifier(new.data)

DexamethasonePredictor(new.data)

MelphalanClassifier(new.data)

MelphalanPredictor(new.data)
}
\arguments{
\item{new.data}{An expression matrix.}

\item{drugs}{An RMA reference object created by rmaPreprocessing.}

\item{cut}{Should the cel files be tested. When set to TRUE bad cel files are automatically discarded.}

\item{type}{For Rituximab, What type of classifier or predictor should be used. Current choices are corrected, uncorrected, lysis, and lysis2.}

\item{calc.cut}{For Rituximab, calculate the cutpoints according to proportions in the data. E.g. calc.cut = c(0.33, 0.66) means that a third is deemed sensitive, intermediate, and resistant respectively.}

\item{cut.spec}{For the lysis type rituximab classifier specify the cut point for unclassified.}

\item{percent.classified}{For the lysis type rituximab classifier specify the percentage of unclassified.}
}
\value{
Expression matrix consisting of normalised array.
}
\description{
Resistance Gene Signatures (REGS) for prediction of resistance to various drugs.
}
\details{
Load cel files into a matrix.
}
\examples{
\donttest{
files <- dir(system.file("extdata/celfiles", package = "hemaClass"), full.names = TRUE)
affyBatch <- readCelfiles(filenames = files)

# The cel files are pre-processed
affyRMA <- rmaPreprocessing(affyBatch)

# the function \\code{rmaPreprocessing} returns median centered and scaled
# expression values in the slot exprs.sc.

# The slot exprs.sc.mean contains mean cetered and scaled expression values.
# This scaling can also be achieved using the function \\code{microarrayScale}.
affyRMA.sc <- microarrayScale(affyRMA$exprs, center = "median")

# We may now use the predictors

# The classifier for Cyclophosphamide, Doxorubicin, and Vincristine:
ResistanceClassifier(affyRMA.sc)

# The predictor for Cyclophosphamide, Doxorubicin, and Vincristine:
ResistancePredictor(affyRMA.sc)

# The classifier for Rituximab into Lysis, Statisk, or Resistant:
RituximabClassifier(affyRMA.sc, type = "lysis2", percent.classified = 100)

# The classifier for Rituximab into Sensitive, Intermediate, or Resistant without
# taking human serum into accout:
RituximabClassifier(affyRMA.sc, type = "uncorrected", calc.cut = c(0.33, 0.66))

# The classifier for Rituximab into Sensitive, Intermediate, or Resistant
# while taking human serum into accout:
RituximabClassifier(affyRMA.sc, type = "corrected")

# For the melphalan classifier we use mean centered and sd scaled expression values:
affyRMA.sc.mean <- microarrayScale(affyRMA$exprs, center = "mean")
MelphalanClassifier(affyRMA.sc.mean)

# For the melphalan predictor we use the original scale:
MelphalanPredictor(affyRMA$exprs)
}
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass.com paper.
}

