% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmaReference}
\alias{rmaReference}
\title{RMA pre-process cel files read into R using readCelfiles according to a reference.}
\usage{
rmaReference(affy.batch, reference, test = FALSE)
}
\arguments{
\item{affy.batch}{An object created by readCelfiles.}

\item{reference}{An RMA reference object created by rmaPreprocessing.}

\item{test}{Should the cel files be tested. When set to TRUE bad cel files are automatically discarded.}
}
\value{
Expression matrix consisting of normalised array.
}
\description{
Pre-process cel files according to the RMA method according to a reference build using rmaPreprocessing.
}
\details{
Load cel files into a matrix.
}
\examples{
files <- dir(system.file("extdata/celfiles", package = "hemaClass"), full.names = TRUE)
affy.batch <- readCelfiles(files)

ref.affy   <- rmaPreprocessing(affy.batch)
ref.affy.2 <- rmaPreprocessing(affy.batch, quantile = ref.affy$quantile)
all(ref.affy.2$exprs - ref.affy$exprs < 0.00000001)

user.affy  <- rmaReference(affy.batch, ref.affy)

all(user.affy$exprs - ref.affy$exprs < 0.00000001)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass.com paper.
}

