% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowSds.R
\name{rowSds}
\alias{colSds}
\alias{colVars}
\alias{rowSds}
\alias{rowVars}
\title{Calculation of standard deviations and variance}
\usage{
rowSds(x, ...)

colSds(x, ...)

rowVars(x, ...)

colVars(x, ...)
}
\arguments{
\item{x}{A \code{matrix}.}

\item{...}{Addtional parameters}
}
\value{
A \code{vector} of variances or standard deviations of each
  row or column of matrix \code{x}.
}
\description{
Calculate standard deviations and variance for each row or column of a
matrix.
}
\examples{
x <- matrix(rnorm(30), 5, 6)
rowSds(x)
colSds(x)
rowVars(x)
colVars(x)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

