% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runHemaClass.R
\name{runHemaClass}
\alias{runHemaClass}
\alias{runhemaclass}
\title{Run the shiny server for hemaClass.}
\usage{
runHemaClass(launch.browser = TRUE, debug.help = FALSE, ...)
}
\arguments{
\item{launch.browser}{logical. If \code{TRUE}, the web browser is used.
Default is \code{TRUE}. If \code{FALSE}, the default from
\code{\link{runApp}} is used.}

\item{debug.help}{logical. Should the code be displayed next to or beneath
the web application. Useful when debugging or showcasing the app.}

\item{\dots}{Arguments passed to \code{\link{runApp}}.}
}
\value{
Starts up a local Shiny server instance of the
  \url{http://hemaclass.org} website for faster analysis.
}
\description{
Run a local version of the server running on \url{http://hemaclass.org}.
Provides an easy-to-used and web-based interface for the \pkg{hemaClass}
package.
}
\note{
If the Rstudio viewer is used, only a single \code{.CEL} file can be
  uploaded, please use your browser instead.
}
\examples{
\dontrun{
runHemaClass()

# For a showcase/debug help mode:
runHemaClass(debug.help = TRUE)

# To use the Rstudio viewer:
runHemaClass(launch.browser = FALSE)
}
}
\author{
Steffen Falgreen Larsen <falgreen (at) me.com>\cr
  Anders Ellern Bilgrau \cr
  Jonas Have
}
\references{
\url{http://hemaclass.org}
}
\seealso{
\code{\link{runApp}}
}

