% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ABCGCB.R
\name{ABCGCB}
\alias{ABCGCB}
\alias{ABCGCBClassifier}
\alias{ABCGCBclassifier}
\title{Classification of cell of origin}
\usage{
ABCGCB(new.data, NC.range = c(0.1, 0.9))
}
\arguments{
\item{new.data}{An expression matrix.}

\item{NC.range}{A numeric vector with values for which the probabilities should be cut.}
}
\value{
Probabilities regarding each patients association with each class.
}
\description{
Classification according to ABC/GCB.
}
\details{
The function ABC/GCB classifies DLBCL patients according to the cell of origin for the tumor.
}
\examples{
\donttest{
files <- dir(system.file("extdata/celfiles", package = "hemaClass"), full.names = TRUE)
affyBatch <- readCelfiles(filenames = files)

# The cel files are pre-processed
affyRMA <- rmaPreprocessing(affyBatch)

# the function \\code{rmaPreprocessing} returns median centered and scaled
# expression values in the slot exprs.sc.

# The slot exprs.sc.mean contains mean cetered and scaled expression values.
# This scaling can also be achieved using the function \\code{microarrayScale}.
affyRMA.sc <- microarrayScale(affyRMA$exprs, center = "median")

# We may now use the ABCGCB classifier
ABCGCB(affyRMA.sc)
}
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the ABC/GCB and hemaClass paper.
}

