% Generated by roxygen2 (4.0.1): do not edit by hand
\name{BAGS}
\alias{BAGS}
\alias{BAGSClassifier}
\alias{BAGSclassifier}
\title{Classification of cell of origin}
\usage{
BAGS(new.data, cut.spec = NULL, percent.classified = 85)
}
\arguments{
\item{new.data}{An expression matrix.}

\item{cut.spec}{A numeric value for which the probabilities should be cut.}

\item{percent.classified}{A numeric value indicating the percentage of patients the should be classified.}
}
\value{
Probabilities regarding each patients association with each class.
}
\description{
B-cell associated gene signatures (BAGS).
}
\details{
The function BAGS classifies DLBCL patients according to the cell of origin for the tumor.
}
\examples{
require(affy)
require(MATT)
u133 <- exprs(readRDS(system.file("extdata/GEPhgu.rda", package = "MATT")))

u133 <- microarrayScale(u133)

BAGS(u133)


huex <- exprs(readRDS(system.file("extdata/GEPexon.full.rda",
                                  package = "MATT")))

u133.conv  <- huex10st_to_u133(huex, method = "weighted", type = "Complex", min.pct = 0.8)
BAGS(u133.conv)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the BAGS and hemaClass paper.
}

