% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAGS.R
\name{BAGS}
\alias{BAGS}
\alias{BAGSClassifier}
\alias{BAGSclassifier}
\title{Classification of cell of origin}
\usage{
BAGS(new.data, cut.spec = NULL, percent.classified = 85)
}
\arguments{
\item{new.data}{An expression \code{matrix}.}

\item{cut.spec}{A \code{numeric} value used to threshold the probabilities 
and determine the class.}

\item{percent.classified}{A \code{numeric} value indicating the percentage of 
patients that should be classified. An alternative to \code{cut.spec}.}
}
\value{
A \code{list} of probabilities regarding each patients association 
  with each class, the determined class, and the used cut-off thresholds.
}
\description{
Classification using the B-cell associated gene signatures (BAGS) classifier 
of Dybkaer et al. [1]. The BAGS classifier attemps to assign the given 
supplied Diffuse Large B-Cell Lymphomas (DLBCL) 
samples to one of five classes according to their resembalace to normal 
Naive (N), Centrocytes (CC), Centroblasts (CB), Memory (M), and 
Plasmablasts (PB) cells.
}
\details{
The function BAGS classifies DLBCL patients according to the cell 
  of origin for the tumor [1].
}
\examples{
\donttest{
files <- dir(system.file("extdata/celfiles", package = "hemaClass"), 
             full.names = TRUE)
affyBatch <- readCelfiles(filenames = files)

# The cel files are pre-processed
affyRMA <- rmaPreprocessing(affyBatch)

# The function rmaPreprocessing returns median centered and scaled
# expression values in the slot exprs.sc.

# The slot exprs.sc.mean contains mean centered and scaled expression values.
# This scaling can also be achieved using the function microarrayScale.
affyRMA.sc <- microarrayScale(affyRMA$exprs, center = "median")

# We may now use the BAGS classifier
BAGS(affyRMA.sc)
}                                                                                                                  
}
\references{
[1] Dybkaer, Karen, Martin Boegsted, Steffen Falgreen, Julie S. Boedker, 
Malene K. Kjeldsen, Alexander Schmitz, Anders E. Bilgrau et al. 
"Diffuse large B-cell lymphoma classification system that associates 
normal B-cell subset phenotypes with prognosis." 
Journal of Clinical Oncology 33, no. 12 (2015): 1379-1388.
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr 
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
