% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemaClass-package.R
\docType{package}
\name{hemaClass-package}
\alias{hemaClass-package}
\alias{HEMACLASS}
\alias{hemaclass}
\title{hemaClass: Classification tool for haematological cancers}
\description{
An R-package for DNA microarray classification.
}
\examples{
# See the following for all exported functions
ls("package:hemaClass")

# To run the hemaClass web-interface:
\dontrun{
require(gdata)
require(survival)
require(shinysky)
runHemaClass()
}
}
\author{
Steffen Falgreen Larsen <sfl (at) rn.dk> \cr
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
