% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAll.R
\name{makeAll}
\alias{makeAll}
\alias{makeNews}
\alias{makeCelfiles}
\alias{makeHowto}
\title{Generate news, celfiles, and help for website}
\usage{
makeAll(makeCelfiles = FALSE)

makeNews(indir = "inst", outdir = "inst/website/www")

makeCelfiles(outdir = "inst/website/www")

makeHowto(indir = "vignettes", outdir = "inst/website/www", ...)
}
\arguments{
\item{makeCelfiles}{logical. Should \code{makeCelfiles()} be executed.
Default \code{FALSE} is faster.}

\item{indir}{The input directory to read the files from.}

\item{outdir}{The output directory to write the files to.}

\item{\dots}{Arguments passed to \code{knitr:knit2html}.}
}
\value{
Writes to files in \code{outdir}. Invisibly returns the path to the 
  written file.
}
\description{
Utility functions for generating the \code{News.html}, \code{celfiles.zip}, 
and \code{howto.html} files used in the website from \code{NEWS.Rd}, 
the celfiles in \code{"inst/extdata/celfiles"}, and
\code{vignettes/howto.Rmd}, respectively. 
Intented primarily for internal use.
}
\note{
The functions assumes by default that the working directory is the
  root of the local \pkg{hemaClass} package. \cr
  The functions overwrites all previously present files.
}
\examples{
\dontrun{
makeAll()

# makeAll is an alias for:
makeNews()
makeCelfiles() 
makeHowto()
}
}
\author{
Anders Ellern Bilgrau
}
