% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ABCGCBCoef.R
\name{readABCGCBCoef}
\alias{readABCGCBCoef}
\title{Read coefficients for the ABC/GCB classifier}
\usage{
readABCGCBCoef()
}
\value{
A matrix with coefficients
}
\description{
Function to read coefficients for the ABC/GCB classifier.
}
\examples{
head(readABCGCBCoef())
}

