% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readBAGSCoef}
\alias{readBAGSCoef}
\title{Read coefficients for the BAGS classifier}
\usage{
readBAGSCoef()
}
\value{
A matrix with coefficients
}
\description{
Function to read coefficients for the BAGS classifier.
}
\examples{
head(readBAGSCoef())
}

