% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readPredCoef}
\alias{readClasCoef}
\alias{readCyclophosphamideClasCoef}
\alias{readCyclophosphamidePredCoef}
\alias{readDoxorubicinClasCoef}
\alias{readDoxorubicinPredCoef}
\alias{readDrugCoef}
\alias{readPredCoef}
\alias{readRituximabClasCoef}
\alias{readRituximabClasCorCoef}
\alias{readRituximabClasLytiskCoef}
\alias{readRituximabPredCoef}
\alias{readRituximabPredCorCoef}
\alias{readVincristineClasCoef}
\alias{readVincristinePredCoef}
\title{Read coefficients for REGS classifiers and predictors}
\usage{
readPredCoef()

readClasCoef()

readCyclophosphamideClasCoef()

readCyclophosphamidePredCoef()

readDoxorubicinClasCoef()

readDoxorubicinPredCoef()

readVincristineClasCoef()

readVincristinePredCoef()

readRituximabClasCoef()

readRituximabPredCoef()

readRituximabClasCorCoef()

readRituximabPredCorCoef()

readRituximabClasLytiskCoef()
}
\value{
A matrix with coefficients
}
\description{
Functions to read coefficients for REGS classifiers and predictors.
}
\examples{
head(readPredCoef())
head(readClasCoef())
head(readCyclophosphamideClasCoef())
head(readCyclophosphamidePredCoef())
head(readDoxorubicinClasCoef())
head(readDoxorubicinPredCoef())
head(readVincristineClasCoef())
head(readVincristinePredCoef())
head(readRituximabClasCoef())
head(readRituximabPredCoef())
head(readRituximabClasCorCoef())
head(readRituximabPredCorCoef())
head(readRituximabClasLytiskCoef())
}

