% Generated by roxygen2 (4.0.1): do not edit by hand
\name{readPredCoef}
\alias{readClasCoef}
\alias{readCyclophosphamideClasCoef}
\alias{readCyclophosphamidePredCoef}
\alias{readDoxorubicinClasCoef}
\alias{readDoxorubicinPredCoef}
\alias{readDrugCoef}
\alias{readPredCoef}
\alias{readVincristineClasCoef}
\alias{readVincristinePredCoef}
\title{Read coefficients for REGS classifiers and predictors}
\usage{
readPredCoef()

readClasCoef()

readCyclophosphamideClasCoef()

readCyclophosphamidePredCoef()

readDoxorubicinClasCoef()

readDoxorubicinPredCoef()

readVincristineClasCoef()

readVincristinePredCoef()
}
\value{
A matrix with coefficients
}
\description{
Functions to read coefficients for REGS classifiers and predictors.
}
\examples{
head(readPredCoef())
head(readClasCoef())
head(readCyclophosphamideClasCoef())
head(readCyclophosphamidePredCoef())
head(readDoxorubicinClasCoef())
head(readDoxorubicinPredCoef())
head(readVincristineClasCoef())
head(readVincristinePredCoef())
}

