% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readReference.R
\name{readLLMPPCHOPreference}
\alias{readCHEPRETROreference}
\alias{readIDRCreference}
\alias{readLLMPPCHOPreference}
\alias{readLLMPPRCHOPreference}
\alias{readMDFCIreference}
\alias{readReference}
\alias{readUAMSreference}
\title{Read references for RMA normalisation based on reference}
\usage{
readLLMPPCHOPreference()

readLLMPPRCHOPreference()

readMDFCIreference()

readIDRCreference()

readCHEPRETROreference()

readUAMSreference()
}
\value{
An object that may be used as an reference.
}
\description{
Functions to read coefficients for REGS classifiers and predictors.
}
\examples{
names(readLLMPPCHOPreference())
names(readLLMPPRCHOPreference())
names(readMDFCIreference())
names(readIDRCreference())
names(readCHEPRETROreference())
names(readUAMSreference())
}

