% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ResistanceClassifier}
\alias{CyclophosphamideClassifier}
\alias{CyclophosphamidePredictor}
\alias{DoxorubicinClassifier}
\alias{DoxorubicinPredictor}
\alias{REGS}
\alias{ResistanceClassifier}
\alias{ResistancePredictor}
\alias{RituximabClassifier}
\alias{RituximabPredictor}
\alias{VincristineClassifier}
\alias{VincristinePredictor}
\title{REGS for prediction of chemoresistance}
\usage{
ResistanceClassifier(new.data, drugs = c("Cyclophosphamide", "Doxorubicin",
  "Vincristine", "Combined"), cut = list(Cyclophosphamide = c(0.455, 0.67),
  Doxorubicin = c(0.1, 0.86), Vincristine = c(0.38, 0.54), Combined = c(0.067,
  0.907)))

ResistancePredictor(new.data, drugs = c("Cyclophosphamide", "Doxorubicin",
  "Vincristine", "Combined"), cut = list(Cyclophosphamide = c(280, 340),
  Doxorubicin = c(280, 320), Vincristine = c(115, 127), Combined = c(200, 295)))

CyclophosphamideClassifier(new.data)

CyclophosphamidePredictor(new.data)

DoxorubicinClassifier(new.data)

DoxorubicinPredictor(new.data)

VincristineClassifier(new.data)

VincristinePredictor(new.data)

RituximabClassifier(new.data, type = "corrected", cut = c(0.33, 0.66),
  calc.cut = NULL, cut.spec = NULL, percent.classified = 85)

RituximabPredictor(new.data, type = "corrected", cut = c(0.33, 0.66),
  calc.cut = NULL)
}
\arguments{
\item{new.data}{An expression matrix.}

\item{drugs}{An RMA reference object created by rmaPreprocessing.}

\item{cut}{Should the cel files be tested. When set to TRUE bad cel files are automatically discarded.}

\item{type}{For Rituximab, What type of classifier or predictor should be used. Current choices are corrected, uncorrected, lysis, and lysis2.}

\item{calc.cut}{For Rituximab, calculate the cutpoints according to proportions in the data. E.g. calc.cut = c(0.33, 0.66) means that a third is deemed sensitive, intermediate, and resistant respectively.}

\item{cut.spec}{For the lysis type rituximab classifier specify the cut point for unclassified.}

\item{percent.classified}{For the lysis type rituximab classifier specify the percentage of unclassified.}
}
\value{
Expression matrix consisting of normalised array.
}
\description{
Resistance Gene Signatures (REGS) for prediction of resistance to various drugs.
}
\details{
Load cel files into a matrix.
}
\examples{
require(affy)
require(MATT)
u133 <- exprs(readRDS(system.file("extdata/GEPhgu.rda", package = "MATT")))

u133 <- microarrayScale(u133)

ResistanceClassifier(u133)
ResistancePredictor(u133)
RituximabClassifier(u133, type = "lysis", percent.classified = 100)
RituximabClassifier(u133, type = "uncorrected", calc.cut = c(0.33, 0.66))
RituximabClassifier(u133, type = "corrected")

huex <- exprs(readRDS(system.file("extdata/GEPexon.full.rda",
                                  package = "MATT")))

u133.conv  <- huex10st_to_u133(huex, method = "weighted", type = "Complex")
ResistanceClassifier(u133.conv)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass.com paper.
}

