% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rmaPreprocessing}
\alias{rmaPreprocessing}
\title{RMA pre-process cel files read into R using readCelfiles}
\usage{
rmaPreprocessing(affy.batch, test = FALSE, quantile = NULL)
}
\arguments{
\item{affy.batch}{An object created by readCelfiles.}

\item{test}{Should the cel files be tested. When set to TRUE bad cel files are automatically discarded.}

\item{quantile}{Specify a vector of quantiles that the data should be normalized to.}
}
\value{
Expression matrix consisting of normalised array.
}
\description{
Pre-process cel files according to the RMA method and create a reference for later use.
}
\details{
Load cel files into a matrix.
}
\examples{
files <- dir(system.file("extdata/celfiles", package = "hemaClass"), full.names = TRUE)
affyBatch <- readCelfiles(filenames = files[1])

affyRMA <- rmaPreprocessing(affyBatch)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass.com paper
}

