% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmaPreprocessing.R
\name{rmaPreprocessing}
\alias{rmaPreprocessing}
\title{RMA pre-process \code{.CEL} files}
\usage{
rmaPreprocessing(affy.batch, test = FALSE, quantile = NULL)
}
\arguments{
\item{affy.batch}{An object created by \code{readCelfiles}.}

\item{test}{Should the \code{.CEL} files be tested. When set to \code{TRUE}
bad \code{.CEL} files are automatically discarded.}

\item{quantile}{Specify a vector of quantiles that the data should be 
normalized to.}
}
\value{
Returns a \code{list} of length 9 with the slots:
  \item{exprs}{The cohort RMA pre-processed \eqn{log_2}-expression 
    \code{matrix}}
  \item{exprs.sc}{As slot \code{$exprs} but scaled to have zero median and 
    unit variance.}
  \item{exprs.sc.mean}{As slot \code{$exprs} but scaled to have zero mean and 
    unit variance.}
  \item{quantile}{The values defining the distribution used in quantile 
    normalization.}
  \item{alpha}{The alpha values of the RMA summarization.}
  \item{sd}{A \code{numeric} of the standard errors of each feature.}
  \item{median}{A \code{numeric} with the medians for each feature.}
  \item{mean}{A \code{numeric} with the means for each feature.}
  \item{bad}{A \code{character} giving the 'bad' arrays. \code{NULL} if no
    bad arrays are found or when \code{test = FALSE}}
}
\description{
Pre-process \code{.CEL} files read into \R using 
\code{\link{readCelfiles}} according to the regular cohort RMA method and 
create a reference for later use.
}
\examples{
# List .CEL files bundled with hemaClass
files <- list.files(system.file("extdata/celfiles", package = "hemaClass"), 
                    full.names = TRUE)
affy.batch <- readCelfiles(filenames = files[1:3]) # Read three first files

# RMA normalize
affy.rma <- rmaPreprocessing(affy.batch) 
str(affy.rma)
}
\references{
\url{http://hemaClass.org}
}
\seealso{
\code{\link{rmaReference}} for one-by-one reference based RMA 
  normalization.
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr 
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
