% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmaReference.R
\name{rmaReference}
\alias{rmaReference}
\title{RMA pre-process \code{.CEL} files read}
\usage{
rmaReference(affy.batch, reference, test = FALSE)
}
\arguments{
\item{affy.batch}{An \code{affy.batch} object, usually created by
\code{\link{readCelfiles}}.}

\item{reference}{An RMA reference object created by
\code{\link{rmaPreprocessing}}.}

\item{test}{Should the supplied \code{.CEL} files be qualtiy tested. When
set to \code{TRUE} bad \code{.CEL} files are automatically discarded.}
}
\value{
Returns a \code{list} of length 3 with the slots:
  \item{exprs}{The one-by-one RMA pre-processed \eqn{log_2}-expression
    \code{matrix}}
  \item{exprs.sc}{As slot \code{$exprs} but scaled to have zero median and
    unit variance.}
  \item{exprs.sc.mean}{As slot \code{$exprs} but scaled to have zero mean and
    unit variance.}
}
\description{
Pre-process \code{.CEL} files according to the one-by-one RMA method
using a reference build using \code{\link{rmaPreprocessing}}.
}
\examples{
# List .CEL files bundled with hemaClass
files <- list.files(system.file("extdata/celfiles", package = "hemaClass"),
                    full.names = TRUE)
affy.batch <- readCelfiles(files) # Read in the .CEL files

# Build references
ref.affy   <- rmaPreprocessing(affy.batch)
ref.affy.2 <- rmaPreprocessing(affy.batch, quantile = ref.affy$quantile)
all(ref.affy.2$exprs - ref.affy$exprs < 0.00000001)

# Or use build-in:
ref.affy <- readCHEPRETROreference()

# RMA one-by-one pre-process using the reference:
user.affy  <- rmaReference(affy.batch, ref.affy)
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr
  Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\references{
\url{http://hemaClass.org}
}
\seealso{
\code{\link{rmaPreprocessing}}, \code{\link{readCelfiles}}
}

