% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rowSds.R
\name{rowSds}
\alias{colSds}
\alias{colVars}
\alias{rowSds}
\alias{rowVars}
\title{Calculation of Standard deveations.}
\usage{
rowSds(x, ...)

colSds(x, ...)

rowVars(x, ...)

colVars(x, ...)
}
\arguments{
\item{x}{A matrix.}

\item{...}{Addtional parameters}
}
\value{
The standard deviation of each row or column of matrix x.
}
\description{
Calculation of Standard deveations.
}
\details{
The function calculates the standard deviation row wise for matrix x.
}
\examples{
2+2
}
\author{
Steffen Falgreen <sfl (at) rn.dk> \cr Anders Ellern Bilgrau <abilgrau
(at) math.aau.dk>
}
\references{
Reference to the hemaClass paper.
}

