% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/runHemaClass.R
\name{runHemaClass}
\alias{runHemaClass}
\title{Run the shiny server for hemaClass.}
\usage{
runHemaClass(Version = 0.1)
}
\arguments{
\item{Version}{What version of hemaClass should be run.}
}
\value{
Starts a shiny server for the analysis
}
\description{
Run a local version of the server running on \code{hemaclass.com}. Provides
an easy web-based interface for the \pkg{hemaClass} package.
}
\examples{
\dontrun{
library("gdata")
library("survival")
library("shinysky")
runHemaClass()
}
}
\author{
Steffen Falgreen Larsen \cr
  Anders Ellern Bilgrau
}

