% Generated by roxygen2 (4.0.1): do not edit by hand
\name{runHemaClass}
\alias{runHemaClass}
\title{Run the shiny server for hemaClass.}
\usage{
runHemaClass()
}
\value{
Starts a shiny server for the analysis
}
\description{
Run the shiny server for hemaClass.
}
\details{
This is a local version of the server running on hemaclass.com
}
\author{
Steffen Falgreen Larsen
  Anders Ellern Bilgrau
}

