% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runHemaClass}
\alias{runHemaClass}
\title{Run the shiny server for hemaClass.}
\usage{
runHemaClass(Version = 0.1)
}
\arguments{
\item{Version}{What version of hemaClass should be run.}
}
\value{
Starts a shiny server for the analysis
}
\description{
Run the shiny server for hemaClass.
}
\details{
This is a local version of the server running on hemaclass.com
}
\examples{
#require(gdata)
#require(survival)
#require(shinysky)
#runHemaClass()
}
\author{
Steffen Falgreen Larsen
  Anders Ellern Bilgrau
}

